Обновления към дата: 11.12.2024  Build Number: 20241211.1

See merge request iara/EAFA-NEW-ISS/131-10.07.2020!165

+++ IARA.Mobile/IARA.Mobile.Insp/IARA.Mobile.Insp/FlyoutPages/Inspections/Dialogs/DeclarationCatchDialog/CatchDialog/CatchDialogViewModel.cs
@@ -73,7 +73,7 @@ namespace IARA.Mobile.Insp.FlyoutPages.Inspections.Dialogs.DeclarationCatchDialo
                 {
                     FishTypeId = FishType.Value,
                     CatchCount = ParseHelper.ParseInteger(CatchCount),
-                    CatchQuantity = ParseHelper.ParseInteger(CatchQuantity),
+                    CatchQuantity = ParseHelper.ParseDecimal(CatchQuantity),
                     PresentationId = Presentation.Value
                 });
             }
